+++ app/Filament/Pages/ArtificialIntelligence.php
@@ -61,21 +61,13 @@ use Illuminate\Validation\Rules\Password;
 /**
  * @property Form $form
  */
-class ArtificialIntelligence extends Page
+class ArtificialIntelligence extends ProfilePage
 {
-    use InteractsWithFormActions;
-
-    protected static string $view = 'filament.pages.profile-save';
-
-    protected static ?string $cluster = ProfileSettings::class;
-
     protected static ?string $slug = 'artificial-intelligence';
 
     protected static ?string $title = 'Artifical Intelligence';
 
     protected static ?int $navigationSort = 40;
-
-    protected static bool $shouldRegisterNavigation = true;
 
     /** @var array<string, mixed> $data */
     public ?array $data = [];
@@ -100,224 +92,4 @@ class ArtificialIntelligence extends Page
             ]);
     }
 
-    public function mount(): void
-    {
-        $this->fillForm();
-    }
-
-    public function getUser(): Authenticatable|Model
-    {
-        $user = Filament::auth()->user();
-
-        if (! $user instanceof Model) {
-            throw new Exception('The authenticated user object must be an Eloquent model to allow the profile page to update it.');
-        }
-
-        return $user;
-    }
-
-    public function save(): void
-    {
-        try {
-            $this->callHook('beforeValidate');
-
-            $data = $this->form->getState();
-
-            $this->callHook('afterValidate');
-
-            $data = $this->mutateFormDataBeforeSave($data);
-
-            $this->callHook('beforeSave');
-
-            $this->handleRecordUpdate($this->getUser(), $data);
-
-            $this->callHook('afterSave');
-        } catch (Halt $exception) {
-            return;
-        }
-
-        if (request()->hasSession() && array_key_exists('password', $data)) {
-            request()->session()->put([
-                'password_hash_' . Filament::getAuthGuard() => $data['password'],
-            ]);
-        }
-
-        $this->data['password'] = null;
-        $this->data['passwordConfirmation'] = null;
-
-        $this->getSavedNotification()?->send();
-
-        $this->dispatch('refresh-branding-bar');
-
-        if ($redirectUrl = $this->getRedirectUrl()) {
-            $this->redirect($redirectUrl);
-        }
-    }
-
-    public function getFormActionsAlignment(): string
-    {
-        return Alignment::Start->value;
-    }
-
-    protected function fillForm(): void
-    {
-        $data = $this->getUser()->attributesToArray();
-
-        $this->callHook('beforeFill');
-
-        $data = $this->mutateFormDataBeforeFill($data);
-
-        $this->form->fill($data);
-
-        $this->callHook('afterFill');
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     *
-     * @return array<string, mixed>
-     */
-    protected function mutateFormDataBeforeFill(array $data): array
-    {
-        return $data;
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     *
-     * @return array<string, mixed>
-     */
-    protected function mutateFormDataBeforeSave(array $data): array
-    {
-        return $data;
-    }
-
-    /**
-     * @param  array<string, mixed>  $data
-     */
-    protected function handleRecordUpdate(Model $record, array $data): Model
-    {
-        $record->update($data);
-
-        return $record;
-    }
-
-    protected function getSavedNotification(): ?Notification
-    {
-        $title = $this->getSavedNotificationTitle();
-
-        if (blank($title)) {
-            return null;
-        }
-
-        return Notification::make()
-            ->success()
-            ->title($this->getSavedNotificationTitle());
-    }
-
-    protected function getSavedNotificationTitle(): ?string
-    {
-        return __('filament-panels::pages/auth/edit-profile.notifications.saved.title');
-    }
-
-    protected function getRedirectUrl(): ?string
-    {
-        return null;
-    }
-
-    protected function getNameFormComponent(): Component
-    {
-        return TextInput::make('name')
-            ->label(__('filament-panels::pages/auth/edit-profile.form.name.label'))
-            ->required()
-            ->maxLength(255)
-            ->autofocus();
-    }
-
-    protected function getEmailFormComponent(): Component
-    {
-        return TextInput::make('email')
-            ->label(__('filament-panels::pages/auth/edit-profile.form.email.label'))
-            ->email()
-            ->required()
-            ->maxLength(255)
-            ->unique(ignoreRecord: true)
-            ->hint(fn (Get $get): string => $get('is_email_visible_on_profile') ? 'Visible on profile' : 'Not visible on profile');
-    }
-
-    protected function getPasswordFormComponent(): Component
-    {
-        return TextInput::make('password')
-            ->label(__('filament-panels::pages/auth/edit-profile.form.password.label'))
-            ->password()
-            ->rule(Password::default())
-            ->autocomplete('new-password')
-            ->dehydrated(fn ($state): bool => filled($state))
-            ->dehydrateStateUsing(fn ($state): string => Hash::make($state))
-            ->live(debounce: 500)
-            ->same('passwordConfirmation');
-    }
-
-    protected function getPasswordConfirmationFormComponent(): Component
-    {
-        return TextInput::make('passwordConfirmation')
-            ->label(__('filament-panels::pages/auth/edit-profile.form.password_confirmation.label'))
-            ->password()
-            ->required()
-            ->visible(fn (Get $get): bool => filled($get('password')))
-            ->dehydrated(false);
-    }
-
-    /**
-     * @return array<int | string, string | Form>
-     */
-    protected function getForms(): array
-    {
-        return [
-            'form' => $this->form(
-                $this->makeForm()
-                    ->schema([
-                        $this->getNameFormComponent(),
-                        $this->getEmailFormComponent(),
-                        $this->getPasswordFormComponent(),
-                        $this->getPasswordConfirmationFormComponent(),
-                    ])
-                    ->operation('edit')
-                    ->model($this->getUser())
-                    ->statePath('data'),
-            ),
-        ];
-    }
-
-    /**
-     * @return array<Action | ActionGroup>
-     */
-    protected function getFormActions(): array
-    {
-        return [
-            $this->getSaveFormAction(),
-            $this->getCancelFormAction(),
-        ];
-    }
-
-    protected function getCancelFormAction(): Action
-    {
-        return Action::make('cancel')
-            ->label(__('filament-panels::pages/auth/edit-profile.actions.cancel.label'))
-            ->url(filament()->getUrl())
-            ->color('gray');
-    }
-
-    protected function getSaveFormAction(): Action
-    {
-        return Action::make('save')
-            ->label('Save')
-            ->submit('save')
-            ->keyBindings(['mod+s']);
-    }
-
-    protected function hasFullWidthFormActions(): bool
-    {
-        return false;
-    }
 }
